requirements.txt
annotated-types==0.7.0
anyio==4.5.0
cachetools==5.5.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
exceptiongroup==1.2.2
fastapi==0.115.0
google-ai-generativelanguage==0.6.9
google-api-core==2.20.0
google-api-python-client==2.146.0
google-auth==2.35.0
google-auth-httplib2==0.2.0
google-generativeai==0.8.1
googleapis-common-protos==1.65.0
greenlet==3.1.0
grpcio==1.66.1
grpcio-status==1.66.1
h11==0.14.0
httplib2==0.22.0
httptools==0.6.1
idna==3.10
Jinja2==3.1.4
joblib==1.4.2
MarkupSafe==2.1.5
nltk==3.9.1
proto-plus==1.24.0
protobuf==5.28.2
psycopg2-binary==2.9.9
pyasn1==0.6.1
pyasn1_modules==0.4.1
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
pyparsing==3.1.4
python-dotenv==1.0.1
python-multipart==0.0.9
regex==2024.9.11
requests==2.32.3
rsa==4.9
sniffio==1.3.1
SQLAlchemy==2.0.35
starlette==0.38.5
tqdm==4.66.5
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.3
uvicorn==0.30.0
uvloop==0.20.0
websockets==13.0.1
